pyproject.toml
[project]
name = "ili-forecast"
version = "0.1.0"
description = "Weekly ILI forecasting for Israel"
authors = [
    {name = "LEAGURARIE", email = "LEA.GURARIE@GMAIL.COM"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=2.0,<4.0",
    "numpy>=1.24,<3.0",
    "matplotlib>=3.7,<4.0",
    "seaborn>=0.12,<1.0",
    "scikit-learn>=1.3,<2.0",
    "statsmodels>=0.14,<1.0",
    "hmmlearn>=0.3,<1.0",
    "pmdarima>=2.0,<3.0",
    "holidays>=0.40,<1.0",
    "pyyaml>=6.0,<7.0",
    "pyarrow>=14.0,<24.0",
    "joblib>=1.3,<2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"